Create Minmax.py
This file is called Minmax since it's ultimately intended to play the game optimally but right now it only contains a function for evaluating game positions that can be accessed by running Game_player.py and typing 'evaluate' into the console after making a move. Note, this code requires a running MySQL server and currently it only works on a standard board and requires the user to have already made the table using workbench but all that will soon be fixed.